requirements.txt
prophet
pandas
psycopg2